Preprocess

+++ utils/GraphUtils.cpp
@@ -113,7 +113,7 @@ void WordPiecePreprocessor::preprocess(ITensor &tensor)
         //const char32_t pad_token[] =   U"[PAD]";
         //const char32_t start_token[] = U"[CLS]";
         //const char32_t end_token[] =   U"[SEP]";
-        preprocess_typed<char32_t,int>(tensor,1);
+        preprocess_typed<char32_t,float16_t>(tensor,1.0);
     }
     else if (tensor.info()->data_type() == DataType::F16)
     {
